Fix a bug that 32-bit total_mem overflows.

+++ hypersh/host/mcount.h
@@ -7,7 +7,7 @@
 #include "common.h"
 
 using std::vector;
-bool hypercall_mcount_init(char *filename, uint32_t total_mem, uint32_t ncores,
+bool hypercall_mcount_init(char *filename, uint64_t total_mem, uint32_t ncores,
                            uint32_t mem_bin, uint32_t acc_bin);
 
 void hypercall_mcount_fini();
